Add NORMAL Dapper Abstraction

+++ Pair.Infrastructure/Entities/Person.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
+using System.Data.SqlTypes;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -12,7 +14,7 @@ namespace Pair.Infrastructure.Entities
 
         public string Name { get; set; } = string.Empty;
 
-        public int Age { get; set; }
+        public byte Age { get; set; }
 
         public string Sex { get; set; } = string.Empty;
     }
